Fixed Error UUID add employee and move models to domain/model

+++ app/src/main/java/com/kotlin/u_park/presentation/screens/auth/RegisterScreen.kt
@@ -165,7 +165,7 @@ fun RegisterScreen(
                             val user = User(
                                 nombre = nombre,
                                 usuario = usuario,
-                                cedula = cedula,
+                                cedula = cedula.toLong(),
                                 telefono = telefono,
                                 correo = correo,
                                 contrasena = contrasena
